Add files via upload
1. Explain the problem in terms of operating system concept? Description
Ans-13
This problem is based on  “barbershop problem”.  The barbershop consists of a waiting room with n chairs, among the chair belongs to the barber and the rest belongs to the customers.  If a customer arrives and there is an empty chair (in waiting room), he sits down, and waits for his turn for a haircut.  If the waiting room is full, the customer simple leaves.  If there are no customers, the barber sleeps.  If a customer arrives, he notifies the barber and waits.  On being notified, the barber wakes up if sleeping, if not sleeping,  the barber keeps cutting.  Note that a barber is notified when a customer signals on the customers semaphore.  
Ans-16
This problem is based on the Scheduler processors. Where we are using Round Robin Concept to Design a scheduling program that is capable of scheduling many processes that comes in at some time interval and are allocated the fixed time is allotted to every process that arrives in the queue. This fixed time is known as time slice or time quantum. The first process that arrives is selected and sent to the processor for execution. If it is not able to complete its execution within the time quantum provided, then an interrupt is generated using an automated timer. The process is then stopped and is sent back at the end of the queue. However, the state is saved and context is thereby stored in memory. This helps the process to resume from the point where it was interrupted.
The scheduler selects another process from the ready queue and dispatches it to the processor for its execution. It is executed until the time Quantum does not exceed. The same steps are repeated until all the process are finished. The round robin algorithm is simple and the overhead in decision making is very low. It is the best scheduling algorithm for achieving better and evenly distributed response time.
2. Write the algorithm for proposed solution of the assigned problem.
Ans-13
   void barber() 
 {                     
 void barber()  
{                      /* barber process */
	while (1)    {
		wait(customers);	 // go to sleep  if of customers is 0        
		wait (mutex);                    // acquire access to critical section                      
		--waiting;	 // decrement count of waiting customers            
		signal(barb-ctrl);	 // invite a customer to get in the barber chair      
		signal(mutex);                   // release the mutex                                               
		cut_hair();                          // cut hair – will take a variable amount of time  
 	}
}

	while (1)    {
		wait(customers);	// go to sleep   if customers is 0         
		wait (mutex);                   // acquire access to critical section                      
		--waiting;	// decrement count of waiting customers             
		signal(barb-ctrl);	// invite a customer to get in the barber chair       
		signal(mutex);                 // release the mutex                                              
		cut_hair();                        //  cut hair – will take a variable amount of time   
 	}
}

4. Explain all the constraints given in the problem. Attach the code snippet of the implemented constraint.
Ans 13
if barber is not sleeping, then barb-ctrl  0, and magnitude of barb-ctrl is number of customers waiting.
In particular, if  barber is not sleeping and barb-ctrl = 0, then barber is cutting hair with no one waiting.
If barber is sleeping, then barb-ctrl = 0  (not cutting & no customers waiting).   
.barb-ctrl may be summarized as follows:
barb-ctrl	barber sleeping	barber not sleeping
0	not cutting hair & no one waiting	cutting hair & no one waiting
< 0  (negative)	cannot happen	cutting hair & customers waiting
Customers block on this barb-ctrl if the barber is busy
The customer semaphore represents the number of people in the waiting room when the barber is cutting hair, and is –1 when the barber is sleeping.  The barber “sleeps” by blocking on the customers semaphore.


5. If you have implemented any additional algorithm to support the solution, explain the need and usage of the same.
Ans-13 &16
no

6. Explain the boundary conditions of the implemented code.
Ans-13
In this program I have applied random function which is used to take the different and unic number to 
Denote the place of visitors as a client.
Ans-16 
In this there are many use of time quantum to solve the round robin problem, without the time quantum we will not be able to solve problem of round robin.